Update for Godot 4.2.2 stable
I've gone ahead and updated this project template for use with Godot 4.2.2. The update includes refactoring basically all of the scripts in the project to make use of new syntax, as well as changing the type on some outdated nodes such as the Raycast nodes and Tween node on the player. I'm no camera expert, but I've tried to emulate the original camera settings from the video as well in the player's Camera3d Physical Attributes section.
Should be ready for more people to try it out!